+++ test/springboot/src-gen/main/java/org/fuin/cqrs4j/springboot/test/model/PersonCreatedEvent.java
@@ -3,8 +3,8 @@ package org.fuin.cqrs4j.springboot.test.model;
 import jakarta.validation.constraints.NotNull;
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonProperty;
-import javax.annotation.Nullable;
 import org.fuin.ddd4j.core.EntityIdPath;
+import org.fuin.ddd4j.core.EventId;
 import org.fuin.ddd4j.core.EventType;
 import org.fuin.ddd4j.jackson.AbstractDomainEvent;
 import org.fuin.esc.api.HasSerializedDataTypeConstant;
@@ -12,8 +12,12 @@ import org.fuin.esc.api.SerializedDataType;
 import org.fuin.objects4j.common.Contract;
 
 import java.io.Serial;
+import java.time.ZonedDateTime;
+import java.util.HashMap;
+import java.util.Map;
 
 import javax.annotation.concurrent.Immutable;
+import org.fuin.utils4j.Utils4J;
 
 
 /**
@@ -46,21 +50,6 @@ public final class PersonCreatedEvent extends AbstractDomainEvent<PersonId> {
         super();
     }
 
-    /**
-     * Constructor with event data.
-     *
-     * @param id TODO Add '@Label' annotation. TODO Add '@Tooltip' annotation.
-     * @param name TODO Add '@Label' annotation. TODO Add '@Tooltip' annotation.
-     */
-    protected PersonCreatedEvent(
-        final PersonId id,
-        final PersonName name
-    ) {
-        super(new EntityIdPath(id));
-        this.id = id;
-        this.name = name;
-    }
-
     @Override
     @JsonIgnore
     public EventType getEventType() {
@@ -72,6 +61,7 @@ public final class PersonCreatedEvent extends AbstractDomainEvent<PersonId> {
      *
      * @return TODO Add '@Label' annotation. TODO Add '@Tooltip' annotation.
      */
+    @NotNull
     public PersonId getId() {
         return id;
     }
@@ -80,13 +70,18 @@ public final class PersonCreatedEvent extends AbstractDomainEvent<PersonId> {
      *
      * @return TODO Add '@Label' annotation. TODO Add '@Tooltip' annotation.
      */
+    @NotNull
     public PersonName getName() {
         return name;
     }
 
     @Override
     public String toString() {
-        return "MyEvent happened";
+        final Map<String, String> vars = new HashMap<>();
+        vars.put("entityIdPath", getEntityIdPath().toString());
+        vars.put("id", "" + id);
+        vars.put("name", "" + name);
+        return Utils4J.replaceVars("MyEvent happened", vars);
     }
 
     /**
@@ -142,6 +137,12 @@ public final class PersonCreatedEvent extends AbstractDomainEvent<PersonId> {
          */
         public PersonCreatedEvent build() {
             ensureBuildableAbstractDomainEvent();
+            if (delegate.getEventId() == null) {
+                this.eventId(new EventId());
+            }
+            if (delegate.getEventTimestamp() == null) {
+                this.timestamp(ZonedDateTime.now());
+            }
             final PersonCreatedEvent result = delegate;
             delegate = new PersonCreatedEvent();
             resetAbstractDomainEvent(delegate);