+++ test/springboot/src/main/java/org/fuin/cqrs4j/springboot/test/app/SpringBootConfig.java
@@ -8,9 +8,6 @@ import com.fasterxml.jackson.databind.SerializationFeature;
 import io.kurrent.dbclient.KurrentDBClient;
 import io.kurrent.dbclient.KurrentDBClientSettings;
 import io.kurrent.dbclient.KurrentDBProjectionManagementClient;
-import org.fuin.cqrs4j.core.SimpleViewRegistry;
-import org.fuin.cqrs4j.core.View;
-import org.fuin.cqrs4j.core.ViewRegistry;
 import org.fuin.cqrs4j.jackson.Cqrs4JacksonModule;
 import org.fuin.cqrs4j.springboot.base.EventstoreConfig;
 import org.fuin.ddd4j.core.EntityIdFactory;
@@ -42,17 +39,11 @@ import java.net.http.HttpClient;
 import java.nio.charset.StandardCharsets;
 import java.time.Duration;
 import java.time.temporal.ChronoUnit;
-import java.util.List;
 
 @Configuration
 public class SpringBootConfig {
 
     private static final Logger LOG = LoggerFactory.getLogger(SpringBootConfig.class);
-
-    @Bean
-    public ViewRegistry viewClassRegistry(List<View> views) {
-        return new SimpleViewRegistry(views);
-    }
 
     @Bean
     public EntityIdFactory entityIdFactory() {