+++ release-notes.md
@@ -6,6 +6,7 @@
     This results in changes in several places.
   - Added new [ViewRegistry](core/src/main/java/org/fuin/cqrs4j/core/ViewRegistry.java) that holds information about available views
   - ([View](core/src/main/java/org/fuin/cqrs4j/core/View.java))s now need to implement a `getBeanName()` method to make it more explicit what it is used for.
+- Added new [springboot-starter](springboot-starter) module for better autoconfiguration
 
 ## 0.6.0
 - Added new Jackson module